Fixes exceptNullValues to not omit fields
Previously, when iterating over the keys of an object fetched by the generated API, keys where the value was `null` were omitted.
Explicitly set the field to `undefined` if the original field value was `null`. This is in line to array elements being set to `undefined` from `null`.
Additionally, typing does not break, since the original object was typed as `{X:string|null}`, the generated type is {X?:string} and so `undefined` is a valid value for that object.